requirements.txt
fastapi
uvicorn
pandas
pydantic
python-multipart
python-docx
